добавил решение по 5 заданию
